Extend block format with health

+++ src/item.h
@@ -1,10 +1,12 @@
 #ifndef _ITEM_H_
 #define _ITEM_H_
 
+#include <cstdint>
+
 namespace konstructs {
     struct ItemStack {
         int amount;
-        int type;
+        uint16_t type;
     };
 };
